chore: Rename files to follow existing naming convention
