+++ apps/ymir-sdl3/CMakeLists.txt
@@ -19,8 +19,8 @@ add_executable(ymir-sdl3
     src/app/settings.hpp
     src/app/shared_context.cpp
     src/app/shared_context.hpp
-    src/app/services/savestates/SaveStateService.cpp
-    src/app/services/savestates/SaveStateService.hpp
+    src/app/services/savestates/save_state_service.cpp
+    src/app/services/savestates/save_state_service.hpp
     src/app/services/savestates/types.cpp
     src/app/services/savestates/types.hpp
     src/app/stb_implementations.cpp
