+++ apps/ymir-sdl3/src/app/app.hpp
@@ -3,7 +3,8 @@
 #include "cmdline_opts.hpp"
 
 #include "shared_context.hpp"
-#include <app/services/savestates/SaveStateService.hpp>
+
+#include "services/savestates/save_state_service.hpp"
 
 #include "ui/windows/about_window.hpp"
 #include "ui/windows/backup_ram_manager_window.hpp"
